app/rsu/requirements.txt
paho-mqtt==1.6.1
flask
flask-socketio
eventlet